Add new and update existing domain entities for Booking, Sales and User related entities
Due to temporary Azure DevOps service outage, my colleague was unable to pull the latest changes from the remote repository.
To maintain progress and ensure the work remains consistent, we practiced Extreme Programming (pair programming) in person.
All changes in this commit were implemented collaboratively, with both of us actively participating in writing and reviewing the code on a shared machine (this machine from where the code was commited).
This approach allowed us to continue development despite the downtime while preserving accurate commit tracking for grading and collaboration history.